Extract trait to its own module
